# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DocPatient Nexus API",
    "version": "2.0.0",
...
    "contact": {
      "name": "Miyaad Shah Joy",
      "email": "docpatientnexus@example.com"
    }
  },
  "servers": [
    {
      "url": "https://docpatient-nexus.onrender.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "schemas": {
...
      "Review": {
        "type": "object",
        "required": [
          "doctor",
          "patient",
          "appointment",
          "review",
          "rating"
        ],
        "properties": {
          "doctor": {
            "type": "string",
            "example": "682787f1fea3f44089558cd6"
          },
          "patient": {
            "type": "string",
            "example": "682787f1fea3f44089558cd7"
          },
          "appointment": {
            "type": "string",
            "example": "682787f1fea3f44089558cd8"
          },
          "review": {
            "type": "string",
            "example": "Dr. Rafiq was very attentive and explained everything clearly. I felt genuinely cared for during my consultation."
          },
          "rating": {
            "type": "number",
            "example": 5
          },
          "reply": {
            "type": "string",
            "example": "Thank you for your kind feedback. We're glad to hear about your experience!"
          },
          "isEdited": {
            "type": "boolean",
            "default": false
          },
          "status": {
            "type": "string",
            "enum": [
              "visible",
              "flagged",
              "hidden"
            ],
            "default": "visible"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-05-16T18:46:09.776Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-05-31T02:45:43.207Z"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "paths": {
...
    "/api/v2/patients/appointments/{appointmentId}/reviews": {
      "post": {
        "tags": [
          "Patients"
        ],
        "summary": "Post a review for an appointment.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Allows a patient to `post a review` for an appointment. The patient must be `logged in` to use this route.",
        "operationId": "postReview",
        "parameters": [
          {
            "name": "appointmentId",
            "in": "path",
            "description": "ID of the appointment to post a review for.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "review": {
                    "type": "string",
                    "example": "Dr. Rafiq was very attentive and explained everything clearly. I felt genuinely cared for during my consultation."
                  },
                  "rating": {
                    "type": "number",
                    "example": 5
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review posted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Review created successfully."
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "review": {
                          "$ref": "#/components/schemas/Review"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "Validation failed."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access. Only logged-in patients can access this route. Log in with a valid `jwt` token.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "You are not authorized to access this route. Please log in."
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden access. Only logged-in patients can access this route.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "You are not allowed to post a review here."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No appointment found with provided ID.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "No appointment found."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error. Something went wrong while processing your request.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Something went wrong. Please try again later."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}